Reorganize package structure
This separates REST data from read only data.
