+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/ApiResult.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.fetcher
+package com.gustavo.competitiveprogrammingapp.cfApi
 
 import com.google.gson.JsonElement
