+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/Fetcher.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.fetcher
+package com.gustavo.competitiveprogrammingapp.cfApi
 
 import com.google.gson.Gson
 import org.slf4j.Logger
