+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/UrlCacheRepository.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.fetcher
+package com.gustavo.competitiveprogrammingapp.cfApi
 
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
