+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/cfApi/resources/ProblemsetProblemsGson.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.fetcher.resources
+package com.gustavo.competitiveprogrammingapp.cfApi.resources
 
 data class ProblemsetProblemsGson(
     val problems: List<Problem>? = null
