--- server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/domain/problem/ProblemStatus.kt
@@ -1,5 +1,0 @@
-package com.gustavo.competitiveprogrammingapp.domain.problem
-
-enum class ProblemStatus {
-    NOTHING, READ, WA, AC
-}
