+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/information/InformationController.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.information
+package com.gustavo.competitiveprogrammingapp.information
 
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
