+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/information/Processor.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.cfApi.information
+package com.gustavo.competitiveprogrammingapp.information
 
 /* Processors update the server database given API structured data */
 interface Processor {
