+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/readOnly/cfContest/CfContestRepository.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.domain.cfContest
+package com.gustavo.competitiveprogrammingapp.readOnly.cfContest
 
 import org.springframework.data.mongodb.repository.MongoRepository
 import org.springframework.stereotype.Repository
