+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/readOnly/cfProblem/CfProblemRepository.kt
@@ -1,7 +1,6 @@
-package com.gustavo.competitiveprogrammingapp.domain.cfProblem
+package com.gustavo.competitiveprogrammingapp.readOnly.cfProblem
 
 import org.springframework.data.mongodb.repository.MongoRepository
-import org.springframework.data.mongodb.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
