+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/rest/problem/EditorialStatus.kt
@@ -1,4 +1,4 @@
-package com.gustavo.competitiveprogrammingapp.domain.problem
+package com.gustavo.competitiveprogrammingapp.rest.problem
 
 enum class EditorialStatus {
     NOTHING, READ_BEFORE_AC, READ_AFTER_AC
