+++ server/src/main/kotlin/com/gustavo/competitiveprogrammingapp/rest/problem/ProblemStatus.kt
@@ -1,0 +1,5 @@
+package com.gustavo.competitiveprogrammingapp.rest.problem
+
+enum class ProblemStatus {
+    NOTHING, READ, WA, AC
+}
